+++ index.ios.js
@@ -1,11 +1,11 @@
 'use strict';
 
 import { AppRegistry } from 'react-native';
-import MatchBreakdown2015 from './app/views/MatchBreakdown2015';
-import MatchBreakdown2016 from './app/views/MatchBreakdown2016';
-import MatchBreakdown2017 from './app/views/MatchBreakdown2017';
-import MatchBreakdown2018 from './app/views/MatchBreakdown2018';
-import MatchBreakdown2019 from './app/views/MatchBreakdown2019';
+import MatchBreakdown2015 from './app/views/breakdowns/MatchBreakdown2015';
+import MatchBreakdown2016 from './app/views/breakdowns/MatchBreakdown2016';
+import MatchBreakdown2017 from './app/views/breakdowns/MatchBreakdown2017';
+import MatchBreakdown2018 from './app/views/breakdowns/MatchBreakdown2018';
+import MatchBreakdown2019 from './app/views/breakdowns/MatchBreakdown2019';
 import EventInsights2016 from './app/views/EventInsights2016';
 import EventInsights2017 from './app/views/EventInsights2017';
 import EventInsights2018 from './app/views/EventInsights2018';
